Moved Anthropic code from prompt.go to anthropic.go
